Agregado de funcion para crear receta.
