+++ menu.py
@@ -4,6 +4,7 @@ import random as r
 import json
 #para crear una visualizacion de una receta especifica.
 from visualizar_RecetaDelDia import VentanaPrincipal as recetaDelDia
+from crear_Receta import App as crearReceta
 
 
 class VentanaPrincipal(ttk.Frame):
@@ -18,7 +19,7 @@ class VentanaPrincipal(ttk.Frame):
         # MENU
         ttk.Button(self.parent, text="Mostrar recetas", command="").grid(row=0, column=2, columnspan=2, padx=10, pady=10, ipadx=14)
         ttk.Button(self.parent, text="Mostrar receta del dia", command=self.recetaDia).grid(row=1, column=2, columnspan=2, padx=10, pady=10, ipadx=14)
-        ttk.Button(self.parent, text="Agregar Receta", command="").grid(row=2, column=2, columnspan=2, padx=10, pady=10, ipadx=15)
+        ttk.Button(self.parent, text="Agregar Receta", command=self.crear_receta).grid(row=2, column=2, columnspan=2, padx=10, pady=10, ipadx=15)
         ttk.Button(self.parent, text="Modificar Receta", command="").grid(row=3, column=2, columnspan=2, padx=10, pady=10, ipadx=10)
         ttk.Button(self.parent, text="Eliminar Receta", command="").grid(row=4, column=2, columnspan=2, padx=10, pady=10, ipadx=14)
         bottonExit = ttk.Button(self.parent, text="Cerrar", command="")
@@ -34,6 +35,10 @@ class VentanaPrincipal(ttk.Frame):
         index = r.randint(0, len(recetas)-1)
         toplevel = tk.Toplevel(self.parent)
         recetaDelDia(toplevel, recetas[index]).grid()
+    
+    def crear_receta(self):
+        toplevel = tk.Toplevel(self.parent)
+        crearReceta(toplevel)
 
 root = tk.Tk()
 VentanaPrincipal(root).grid()
